requirements.txt
selenium==4.18.1
pytest==8.0.2
allure-pytest==2.13.2